[migration_Kotilin] 소스 코드 리팩토링

+++ src/main/kotlin/com/example/board/repository/BoardRepository.kt
@@ -9,5 +9,5 @@ interface BoardRepository : JpaRepository<Board, Long> {
 
     fun findTop4ByDelYnOrderByCreateDateDesc(delYn: Int) : MutableList<Board>
     fun findByDelYn(delYn : Int, pageable : Pageable) : Page<Board>
-    fun findByIdAndDelYn(id : Int, delYn : Int) :Board
+    fun findByIdAndDelYn(id: Long, delYn: Int) :Board
 }
